+++ src/app/shared/components/dialog-switcher-button/dialog-switcher-button.component.spec.ts
@@ -8,7 +8,7 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { MatDialog, MatDialogModule } from '@angular/material/dialog';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { InformationDialogComponent } from 'src/app/shared/dialogs/information-dialog/information-dialog.component';
+import { InformationDialogComponent } from 'src/app/shared/dialogs/information/information-dialog.component';
 import { DialogSwitcherButtonComponent } from './dialog-switcher-button.component';
 
 describe('DialogSwitcherButtonComponent', () => {
