+++ src/app/shared/dialogs/information/InformationDialogData.ts
