+++ src/app/shared/dialogs/information/information-dialog.component.css
